+++ src/app/views/NotificationSection.svelte
@@ -1,14 +1,13 @@
 <script lang="ts">
   import PersonBadge from "src/app/shared/PersonBadge.svelte"
 
+  export let label
   export let pubkeys
 </script>
 
-<div class="flex flex-col gap-2">
-  <slot />
-  <div class="flex flex-col gap-2">
-    {#each pubkeys as pubkey}
-      <PersonBadge {pubkey} />
-    {/each}
-  </div>
+<p class="text-xl">{label}</p>
+<div class="flex flex-col gap-4">
+  {#each pubkeys as pubkey}
+    <PersonBadge {pubkey} />
+  {/each}
 </div>